Conflicts:
	includes/feeds.php

+++ includes/admin.php
@@ -16,6 +16,13 @@ class WP_Stream_Admin {
 	 */
 	public static $list_table = null;
 
+	/**
+	 * Option to disable access to Stream
+	 *
+	 * @var bool
+	 */
+	public static $disable_access = false;
+
 	const ADMIN_BODY_CLASS    = 'wp_stream_screen';
 	const RECORDS_PAGE_SLUG   = 'wp_stream';
 	const SETTINGS_PAGE_SLUG  = 'wp_stream_settings';
@@ -29,14 +36,20 @@ class WP_Stream_Admin {
 		add_filter( 'user_has_cap', array( __CLASS__, '_filter_user_caps' ), 10, 4 );
 		add_filter( 'role_has_cap', array( __CLASS__, '_filter_role_caps' ), 10, 3 );
 
-		// Add admin body class
-		add_filter( 'admin_body_class', array( __CLASS__, 'admin_body_class' ) );
+		self::$disable_access = apply_filters( 'wp_stream_disable_admin_access', false );
 
 		// Register settings page
 		add_action( 'admin_menu', array( __CLASS__, 'register_menu' ) );
 
+		// Admin notices
+		add_action( 'admin_notices', array( __CLASS__, 'admin_notices' ) );
+
+		// Add admin body class
+		add_filter( 'admin_body_class', array( __CLASS__, 'admin_body_class' ) );
+
 		// Plugin action links
 		add_filter( 'plugin_action_links', array( __CLASS__, 'plugin_action_links' ), 10, 2 );
+		add_filter( 'network_admin_plugin_action_links', array( __CLASS__, 'plugin_action_links' ), 10, 2 );
 
 		// Load admin scripts and styles
 		add_action( 'admin_enqueue_scripts', array( __CLASS__, 'admin_enqueue_scripts' ) );
@@ -44,6 +57,9 @@ class WP_Stream_Admin {
 
 		// Reset Streams database
 		add_action( 'wp_ajax_wp_stream_reset', array( __CLASS__, 'wp_ajax_reset' ) );
+
+		// Reset Streams settings
+		add_action( 'wp_ajax_wp_stream_defaults', array( __CLASS__, 'wp_ajax_defaults' ) );
 
 		// Uninstall Streams and Deactivate plugin
 		add_action( 'wp_ajax_wp_stream_uninstall', array( __CLASS__, 'uninstall_plugin' ) );
@@ -78,6 +94,9 @@ class WP_Stream_Admin {
 			case 'data_erased':
 				printf( '<div class="updated"><p>%s</p></div>', __( 'All records have been successfully erased.', 'stream' ) );
 				break;
+			case 'settings_reset':
+				printf( '<div class="updated"><p>%s</p></div>', __( 'All site settings have been successfully reset.', 'stream' ) );
+				break;
 		}
 	}
 
@@ -85,9 +104,17 @@ class WP_Stream_Admin {
 	 * Register menu page
 	 *
 	 * @action admin_menu
-	 * @return void
+	 * @return bool|void
 	 */
 	public static function register_menu() {
+		if ( is_network_admin() && ! is_plugin_active_for_network( WP_STREAM_PLUGIN ) ) {
+			return false;
+		}
+
+		if ( self::$disable_access ) {
+			return false;
+		}
+
 		self::$screen_id['main'] = add_menu_page(
 			__( 'Stream', 'stream' ),
 			__( 'Stream', 'stream' ),
@@ -107,8 +134,12 @@ class WP_Stream_Admin {
 			array( __CLASS__, 'render_page' )
 		);
 
-		// Register the list table early, so it associates the column headers with 'Screen settings'
-		add_action( 'load-' . self::$screen_id['main'], array( __CLASS__, 'register_list_table' ) );
+		do_action( 'wp_stream_admin_menu_screens' );
+
+		if ( ! WP_Stream_Install::$update_required ) {
+			// Register the list table early, so it associates the column headers with 'Screen settings'
+			add_action( 'load-' . self::$screen_id['main'], array( __CLASS__, 'register_list_table' ) );
+		}
 	}
 
 	/**
@@ -123,10 +154,11 @@ class WP_Stream_Admin {
 	public static function admin_enqueue_scripts( $hook ) {
 		wp_register_script( 'select2', WP_STREAM_URL . 'ui/select2/select2.min.js', array( 'jquery' ), '3.4.5', true );
 		wp_register_style( 'select2', WP_STREAM_URL . 'ui/select2/select2.css', array(), '3.4.5' );
-
 		wp_register_script( 'timeago', WP_STREAM_URL . 'ui/timeago/timeago.js', array(), '0.2.0', true );
+
 		$locale    = substr( get_locale(), 0, 2 );
 		$file_tmpl = 'ui/timeago/locale/jquery.timeago.%s.js';
+
 		if ( file_exists( WP_STREAM_DIR . sprintf( $file_tmpl, $locale ) ) ) {
 			wp_register_script( 'timeago-locale', WP_STREAM_URL . sprintf( $file_tmpl, $locale ), array( 'timeago' ), '1' );
 		} else {
@@ -153,14 +185,15 @@ class WP_Stream_Admin {
 				array(
 					'i18n'            => array(
 						'confirm_purge'     => __( 'Are you sure you want to delete all Stream activity records from the database? This cannot be undone.', 'stream' ),
+						'confirm_defaults'  => __( 'Are you sure you want to reset all site settings to default? This cannot be undone.', 'stream' ),
 						'confirm_uninstall' => __( 'Are you sure you want to uninstall and deactivate Stream? This will delete all Stream tables from the database and cannot be undone.', 'stream' ),
 					),
-					'gmt_offset'      => get_option( 'gmt_offset' ),
-					'current_screen'  => $hook,
-					'current_page'    => isset( $_GET['paged'] ) ? esc_js( $_GET['paged'] ) : '1',
-					'current_order'   => isset( $_GET['order'] ) ? esc_js( $_GET['order'] ) : 'desc',
-					'current_query'   => json_encode( $_GET ),
-					'filter_controls' => get_user_meta( get_current_user_id(), 'stream_toggle_filters', true ),
+					'gmt_offset'     => get_option( 'gmt_offset' ),
+					'current_screen' => $hook,
+					'current_page'   => isset( $_GET['paged'] ) ? esc_js( $_GET['paged'] ) : '1',
+					'current_order'  => isset( $_GET['order'] ) ? esc_js( $_GET['order'] ) : 'desc',
+					'current_query'  => json_encode( $_GET ),
+					'filters'        => self::$list_table ? self::$list_table->get_filters() : false,
 				)
 			);
 		}
@@ -186,8 +219,9 @@ class WP_Stream_Admin {
 	/**
 	 * Add menu styles for various WP Admin skins
 	 *
+	 * @uses wp_add_inline_style()
 	 * @action admin_enqueue_scripts
-	 * @return void
+	 * @return bool true on success false on failure
 	 */
 	public static function admin_menu_css() {
 		wp_register_style( 'jquery-ui', '//ajax.googleapis.com/ajax/libs/jqueryui/1.10.1/themes/base/jquery-ui.css', array(), '1.10.1' );
@@ -256,10 +290,16 @@ class WP_Stream_Admin {
 	 */
 	public static function plugin_action_links( $links, $file ) {
 		if ( plugin_basename( WP_STREAM_DIR . 'stream.php' ) === $file ) {
-			$admin_page_url = add_query_arg( array( 'page' => self::SETTINGS_PAGE_SLUG ), admin_url( self::ADMIN_PARENT_PAGE ) );
+
+			// Don't show links in Network Admin if Stream isn't network enabled
+			if ( is_network_admin() && is_multisite() && ! is_plugin_active_for_network( WP_STREAM_PLUGIN ) ) {
+				return $links;
+			}
+
+			$admin_page_url = add_query_arg( array( 'page' => self::SETTINGS_PAGE_SLUG ), is_network_admin() ? network_admin_url( self::ADMIN_PARENT_PAGE ) : admin_url( self::ADMIN_PARENT_PAGE ) );
 			$links[] = sprintf( '<a href="%s">%s</a>', esc_url( $admin_page_url ), esc_html__( 'Settings', 'stream' ) );
 
-			$url     = add_query_arg(
+			$url = add_query_arg(
 				array(
 					'action'          => 'wp_stream_uninstall',
 					'wp_stream_nonce' => wp_create_nonce( 'stream_nonce' ),
@@ -318,43 +358,62 @@ class WP_Stream_Admin {
 	 * @return void
 	 */
 	public static function render_page() {
+
+		$option_key  = WP_Stream_Settings::$option_key;
+		$form_action = apply_filters( 'wp_stream_settings_form_action', admin_url( 'options.php' ) );
+
+		$page_title       = apply_filters( 'wp_stream_settings_form_title', get_admin_page_title() );
+		$page_description = apply_filters( 'wp_stream_settings_form_description', '' );
+
+		$sections   = WP_Stream_Settings::get_fields();
+		$active_tab = wp_stream_filter_input( INPUT_GET, 'tab' );
+
+		if ( WP_Stream_Install::$update_required ) {
+			printf( '<div class="wrap"><h2>%s</h2></div>', $page_title );
+			return;
+		}
+
 		?>
 		<div class="wrap">
 
-			<h2><?php _e( 'Stream Settings', 'stream' ) ?></h2>
+			<h2><?php echo esc_html( $page_title ); ?></h2>
+
+			<?php if ( ! empty( $page_description ) ) : ?>
+				<p><?php echo esc_html( $page_description ); ?></p>
+			<?php endif; ?>
+
 			<?php settings_errors() ?>
 
-			<?php
-			$sections   = WP_Stream_Settings::get_fields();
-			$active_tab = wp_stream_filter_input( INPUT_GET, 'tab' );
-			?>
-
-			<h2 class="nav-tab-wrapper">
-				<?php $i = 0 ?>
-				<?php foreach ( $sections as $section => $data ) : ?>
-					<?php $i ++ ?>
-					<?php $is_active = ( ( 1 === $i && ! $active_tab ) || $active_tab === $section ) ?>
-					<a href="<?php echo esc_url( add_query_arg( 'tab', $section ) ) ?>" class="nav-tab<?php if ( $is_active ) { echo esc_attr( ' nav-tab-active' ); } ?>">
-						<?php echo esc_html( $data['title'] ) ?>
-					</a>
-				<?php endforeach; ?>
-			</h2>
+			<?php if ( count( $sections ) > 1 ) : ?>
+				<h2 class="nav-tab-wrapper">
+					<?php $i = 0 ?>
+					<?php foreach ( $sections as $section => $data ) : ?>
+						<?php $i ++ ?>
+						<?php $is_active = ( ( 1 === $i && ! $active_tab ) || $active_tab === $section ) ?>
+						<a href="<?php echo esc_url( add_query_arg( 'tab', $section ) ) ?>" class="nav-tab<?php if ( $is_active ) { echo esc_attr( ' nav-tab-active' ); } ?>">
+							<?php echo esc_html( $data['title'] ) ?>
+						</a>
+					<?php endforeach; ?>
+				</h2>
+			<?php endif; ?>
 
 			<div class="nav-tab-content" id="tab-content-settings">
-				<form method="post" action="options.php" enctype="multipart/form-data">
-					<?php
-					$i = 0;
+
+				<form method="post" action="<?php echo esc_attr( $form_action ) ?>" enctype="multipart/form-data">
+		<?php
+		$i = 0;
 		foreach ( $sections as $section => $data ) {
-						$i++;
-						$is_active = ( ( 1 === $i && ! $active_tab ) || $active_tab === $section );
+			$i++;
+			$is_active = ( ( 1 === $i && ! $active_tab ) || $active_tab === $section );
 			if ( $is_active ) {
-							settings_fields( WP_Stream_Settings::KEY );
-							do_settings_sections( WP_Stream_Settings::KEY );
-						}
-					}
-					submit_button();
-					?>
+				settings_fields( $option_key );
+				do_settings_sections( $option_key );
+			}
+		}
+		submit_button();
+		?>
 				</form>
+
 			</div>
 
 		</div>
@@ -367,10 +426,22 @@ class WP_Stream_Admin {
 	}
 
 	public static function stream_page() {
-		self::$list_table->prepare_items();
+		$page_title = __( 'Stream Records', 'stream' );
 
 		echo '<div class="wrap">';
-		printf( '<h2>%s</h2>', __( 'Stream Records', 'stream' ) ); // xss ok
+
+		if ( is_network_admin() ) {
+			$site_count = sprintf( _n( '1 site', '%d sites', get_blog_count(), 'stream' ), get_blog_count() );
+			printf( '<h2>%s (%s)</h2>', __( 'Stream Records', 'stream' ), $site_count ); // xss ok
+		} else {
+			printf( '<h2>%s</h2>', __( 'Stream Records', 'stream' ) ); // xss ok
+		}
+
+		if ( WP_Stream_Install::$update_required ) {
+			return;
+		}
+
+		self::$list_table->prepare_items();
 		self::$list_table->display();
 		echo '</div>';
 	}
@@ -383,10 +454,10 @@ class WP_Stream_Admin {
 			wp_redirect(
 				add_query_arg(
 					array(
-						'page'    => 'wp_stream_settings',
+						'page'    => is_network_admin() ? 'wp_stream_network_settings' : 'wp_stream_settings',
 						'message' => 'data_erased',
 					),
-					admin_url( self::ADMIN_PARENT_PAGE )
+					is_plugin_active_for_network( WP_STREAM_PLUGIN ) ? network_admin_url( self::ADMIN_PARENT_PAGE ) : admin_url( self::ADMIN_PARENT_PAGE )
 				)
 			);
 			exit;
@@ -395,8 +466,13 @@ class WP_Stream_Admin {
 		}
 	}
 
-	public static function erase_stream_records() {
+	private static function erase_stream_records() {
 		global $wpdb;
+
+		$where = '';
+		if ( is_multisite() && ! is_plugin_active_for_network( WP_STREAM_PLUGIN ) ) {
+			$where .= $wpdb->prepare( ' AND `blog_id` = %d', get_current_blog_id() );
+		}
 
 		$wpdb->query(
 			$wpdb->prepare(
@@ -406,10 +482,49 @@ class WP_Stream_Admin {
 				ON `context`.`record_id` = `stream`.`ID`
 				LEFT JOIN {$wpdb->streammeta} AS `meta`
 				ON `meta`.`record_id` = `stream`.`ID`
-				WHERE `stream`.`type` = %s;",
+				WHERE `stream`.`type` = %s
+				$where;",
 				'stream'
 			)
 		);
+	}
+
+	public static function wp_ajax_defaults() {
+		check_ajax_referer( 'stream_nonce', 'wp_stream_nonce' );
+
+		if ( ! is_plugin_active_for_network( WP_STREAM_PLUGIN ) ) {
+			wp_die( "You don't have sufficient privileges to do this action." );
+		}
+
+		if ( current_user_can( self::SETTINGS_CAP ) ) {
+			self::reset_stream_settings();
+			wp_redirect(
+				add_query_arg(
+					array(
+						'page'    => is_network_admin() ? 'wp_stream_network_settings' : 'wp_stream_settings',
+						'message' => 'settings_reset',
+					),
+					is_plugin_active_for_network( WP_STREAM_PLUGIN ) ? network_admin_url( self::ADMIN_PARENT_PAGE ) : admin_url( self::ADMIN_PARENT_PAGE )
+				)
+			);
+			exit;
+		} else {
+			wp_die( "You don't have sufficient privileges to do this action." );
+		}
+	}
+
+	private static function reset_stream_settings() {
+		global $wpdb;
+
+		$blogs = wp_get_sites();
+
+		if ( $blogs ) {
+			foreach ( $blogs as $blog ) {
+				switch_to_blog( $blog['blog_id'] );
+				delete_option( WP_Stream_Settings::KEY );
+			}
+			restore_current_blog();
+		}
 	}
 
 	/**
@@ -425,30 +540,52 @@ class WP_Stream_Admin {
 			// Prevent stream action from being fired on plugin
 			remove_action( 'deactivate_plugin', array( 'WP_Stream_Connector_Installer', 'callback' ), null );
 
-			// Deactivate the plugin
-			deactivate_plugins( plugin_basename( WP_STREAM_DIR ) . '/stream.php' );
+			// Plugin is being uninstalled from only one of the multisite blogs
+			if ( is_multisite() && ! is_plugin_active_for_network( WP_STREAM_PLUGIN ) ) {
+				$blog_id = get_current_blog_id();
+
+				$wpdb->query( "DELETE FROM {$wpdb->base_prefix}stream WHERE blog_id = $blog_id" );
+
+				delete_option( plugin_basename( WP_STREAM_DIR ) . '_db' );
+				delete_option( WP_Stream_Settings::KEY );
+			} else {
+				// Delete all tables
+				foreach ( WP_Stream_DB::get_instance()->get_table_names() as $table ) {
+					$wpdb->query( "DROP TABLE $table" );
+				}
+
+				// Delete database options
+				if ( is_multisite() ) {
+					$blogs = wp_get_sites();
+					foreach ( $blogs as $blog ) {
+						switch_to_blog( $blog['blog_id'] );
+						delete_option( plugin_basename( WP_STREAM_DIR ) . '_db' );
+						delete_option( WP_Stream_Settings::KEY );
+					}
+					restore_current_blog();
+				}
+
+				// Delete database option
+				delete_site_option( plugin_basename( WP_STREAM_DIR ) . '_db' );
+				delete_site_option( WP_Stream_Settings::KEY );
+				delete_site_option( WP_Stream_Settings::DEFAULTS_KEY );
+				delete_site_option( WP_Stream_Settings::NETWORK_KEY );
+				delete_site_option( 'dashboard_stream_activity_options' );
+			}
 
 			// Delete scheduled cron event hooks
 			wp_clear_scheduled_hook( 'stream_auto_purge' ); // Deprecated hook
 			wp_clear_scheduled_hook( 'wp_stream_auto_purge' );
 
-			// Delete all tables
-			foreach ( WP_Stream_DB::get_instance()->get_table_names() as $table ) {
-				$wpdb->query( "DROP TABLE $table" );
-			}
-
-			// Delete database option
-			delete_option( WP_Stream_Install::KEY );
-			delete_option( WP_Stream_Settings::KEY );
-			delete_option( 'dashboard_stream_activity_options' );
+			// Deactivate the plugin
+			deactivate_plugins( plugin_basename( WP_STREAM_DIR ) . '/stream.php' );
 
 			// Redirect to plugin page
-			wp_redirect( add_query_arg( array( 'deactivate' => true ), admin_url( 'plugins.php' ) ) );
+			wp_redirect( add_query_arg( array( 'deactivate' => true ), is_network_admin() ? network_admin_url( 'plugins.php' ) : admin_url( 'plugins.php' ) ) );
 			exit;
 		} else {
 			wp_die( "You don't have sufficient privileges to do this action." );
 		}
-
 	}
 
 	public static function purge_schedule_setup() {
@@ -460,11 +597,27 @@ class WP_Stream_Admin {
 	public static function purge_scheduled_action() {
 		global $wpdb;
 
-		$options = WP_Stream_Settings::get_options();
-		$days    = $options['general_records_ttl'];
-		$date    = new DateTime( 'now', $timezone = new DateTimeZone( 'UTC' ) );
+		// Don't purge if in Network Admin if Stream isn't network enabled
+		if ( is_network_admin() && is_multisite() && ! is_plugin_active_for_network( WP_STREAM_PLUGIN ) ) {
+			return;
+		}
+
+		if ( is_multisite() && is_plugin_active_for_network( WP_STREAM_PLUGIN ) ) {
+			$options = (array) get_site_option( WP_Stream_Settings::NETWORK_KEY, array() );
+		} else {
+			$options = WP_Stream_Settings::get_options();
+		}
+
+		$days = $options['general_records_ttl'];
+		$date = new DateTime( 'now', $timezone = new DateTimeZone( 'UTC' ) );
 
 		$date->sub( DateInterval::createFromDateString( "$days days" ) );
+
+		$where = $wpdb->prepare( ' AND `stream`.`created` < %s', $date->format( 'Y-m-d H:i:s' ) );
+
+		if ( is_multisite() && ! is_plugin_active_for_network( WP_STREAM_PLUGIN ) ) {
+			$where .= $wpdb->prepare( ' AND `blog_id` = %d', get_current_blog_id() );
+		}
 
 		$wpdb->query(
 			$wpdb->prepare(
@@ -475,7 +628,7 @@ class WP_Stream_Admin {
 				LEFT JOIN {$wpdb->streammeta} AS `meta`
 				ON `meta`.`record_id` = `stream`.`ID`
 				WHERE `stream`.`type` = %s
-				AND `stream`.`created` < %s;",
+				$where;",
 				'stream',
 				$date->format( 'Y-m-d H:i:s' )
 			)
